Fix bugs, add 8x8 unroll

+++ m1_computer_systems/optimization/optimization-prework/vec/vec.h
@@ -1,5 +1,5 @@
-// typedef float data_t;
-typedef long data_t;
+typedef float data_t;
+// typedef long data_t;
 
 /* Create abstract data type for vector */
 typedef struct {
